pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faf-sdk"
version = "1.0.0"
description = "Python SDK for FAF (Foundational AI-context Format) - IANA-registered application/vnd.faf+yaml"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "wolfejam", email = "wolfejam@faf.one"}
]
keywords = [
    "faf",
    "ai-context",
    "yaml",
    "mcp",
    "claude",
    "grok",
    "project-context"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup"
]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0"
]

[project.urls]
Homepage = "https://faf.one"
Documentation = "https://github.com/Wolfe-Jam/faf-python-sdk"
Repository = "https://github.com/Wolfe-Jam/faf-python-sdk"
Issues = "https://github.com/Wolfe-Jam/faf-python-sdk/issues"

[tool.hatch.build.targets.wheel]
packages = ["faf_sdk"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
